Update Decision Trees of Patients Treatment.ipynb
In this situation, the patients who have the same illness will be treated with different drugs (Drug A, Drug B, Drug c, Drug x and y) depending on their attributes. The features of this dataset are Age, Sex, Blood Pressure, and the Cholesterol of the patients, and the target is the drug that each patient responded to.